requirements.txt
anyio
starlette
pydantic
fastapi
python-multipart
sqlalchemy
uvicorn
typing-extensions
psycopg2
websockets
passlib
PyJWT
python-jose
 

